Refactor: Componentize QueryEditor
